+++ test/ua_inspector/plug_test.exs
@@ -30,7 +30,7 @@ defmodule UAInspector.PlugTest do
     conn = conn(:get, "/") |> Router.call(@opts)
 
     assert 404 == conn.status
-    assert %UAInspector.Result{user_agent: nil} == UAInspector.Plug.get_result(conn)
+    assert %UAInspector.Result{user_agent: ""} == UAInspector.Plug.get_result(conn)
   end
 
   test "empty result for empty agent" do
@@ -86,6 +86,37 @@ defmodule UAInspector.PlugTest do
              },
              os: %UAInspector.Result.OS{name: "iOS", platform: :unknown, version: "7.0.4"},
              os_family: "iOS",
+             user_agent: ^agent
+           } = UAInspector.Plug.get_result(conn)
+  end
+
+  test "result with client hints" do
+    agent = "client hint only"
+
+    conn =
+      conn(:get, "/")
+      |> put_req_header("user-agent", agent)
+      |> put_req_header("sec-ch-ua", "\"Chromium\";v=\"109\", \"Not_A Brand\";v=\"99\"")
+      |> put_req_header("sec-ch-ua-mobile", "?0")
+      |> put_req_header("sec-ch-ua-platform", "\"Linux\"")
+      |> put_req_header("sec-fetch-dest", "document")
+      |> put_req_header("sec-fetch-mode", "navigate")
+      |> put_req_header("sec-fetch-site", "none")
+      |> put_req_header("sec-fetch-user", "?1")
+      |> Router.call(@opts)
+
+    assert 200 == conn.status
+
+    assert %UAInspector.Result{
+             browser_family: "Chrome",
+             client: %UAInspector.Result.Client{
+               name: "Chromium",
+               type: "browser",
+               version: "109"
+             },
+             device: %UAInspector.Result.Device{type: "desktop"},
+             os: %UAInspector.Result.OS{name: "GNU/Linux"},
+             os_family: "GNU/Linux",
              user_agent: ^agent
            } = UAInspector.Plug.get_result(conn)
   end